50_check_unity_support: only run when called from an unity session
The content of $XDG_CURRENT_DESKTOP is checked inside a sub-shell
not to affect the value of  IFS for all the scripts sourced by the X session.

If running in Unity we should call unity_support_test, not inside gnome
ubuntu X sessions.

XDG_CURRENT_DESKTOP values are currently the followings:
 - unity before 17.10: Unity
 - unity after 17.10: Unity:Unity7:ubuntu
 - gnome session: GNOME
 - ubuntu gnome session: ubuntu:GNOME