Revert: "debian: don't consider the Xsession.d script as a config file anymore"
It seems something there's not full agreement on it, so to avoid further discussions
I'm reverting this back. However I strongly believe that given the current tools we have
this is the way to go, as debian doesn't provide other tools to get this done.
Nor X11 can move files outside of /etc.